A domain independent recommender built on pySpark
This is a recommendation engine I built using for a project at
University of Michigan's School of information.
It uses Spark's Python API and some code from Spark 1.2 Machine Learning
Library.
The idea is that it would be domain independent recommendation engine
focusing on assessing the users culture instead of their specific,
product dependent tastes. It does this my simulating transmission of
memes (memes from Dawkins not the LOL cats).  The other consideration I
took to make it domain independent is by using vector space distance
verses categories by their nature encode some cultural bias and is
against what I am attempting to do.
At this point I get rankings that are inconsistent with the training
set. One area that needs to be worked on is dealing with terms that are
context independent and where the value can change based on modifiers.
ie if Some ranks a place lowly and says their is horrible service it
means that service is important where as if they say there was a
horrible burrito with out any other knowledge you could say they maybe
aren't the biggest fan of burritos (may not be true but isn't self
evidently false)